feat: move rooms suits to component
