+++ src/pages/index.js
@@ -1,9 +1,14 @@
 import Homepage from "./homepage";
+import Navbar from "@/component/navbar"
+import Footer from "@/component/footer";
+
 
 export default function Home() {
   return (
     <div>
+      <Navbar/>
       <Homepage />
+      <Footer />
     </div>
   );
 }
